+++ tasks/accounts.ts
@@ -1,0 +1,11 @@
+import { task } from "hardhat/config";
+
+// This is a sample Hardhat task. To learn how to create your own go to
+// https://hardhat.org/guides/create-task.html
+task("accounts", "Prints the list of accounts", async (args, { ethers }) => {
+  const accounts = await ethers.getSigners();
+
+  for (const account of accounts) {
+    console.log(await account.address);
+  }
+});
