+++ tasks/index.ts
@@ -1,0 +1,7 @@
+import "./accounts";
+import "./add-liquidity";
+import "./cpp-address";
+import "./cpp-deploy";
+import "./erc20-approve";
+import "./strategy";
+import "./whitelist";
